+++ src/main.tsx
@@ -5,20 +5,9 @@ import { ChatbotProvider } from "@/context/ChatbotContext.tsx";
 
 import App from "./App.tsx";
 
-const chatbotWrapper = document.createElement("div");
-chatbotWrapper.setAttribute("id", "rs-chatbot-wrapper");
+import { createDOMElements } from "./utils/createDOMElements.ts";
 
-const root = document.createElement("div");
-root.setAttribute("id", "rs-chatbot-root");
-
-const chatbotRecaptcha = document.createElement("div");
-chatbotRecaptcha.setAttribute("id", "rs-chatbot-recaptcha");
-chatbotRecaptcha.style.display = "none";
-
-chatbotWrapper.appendChild(root);
-chatbotWrapper.appendChild(chatbotRecaptcha);
-
-document.body.appendChild(chatbotWrapper);
+createDOMElements();
 
 ReactDOM.createRoot(document.querySelector("#rs-chatbot-root")!).render(
   <React.StrictMode>
